Change passingScore to a float

+++ GeneticProgramming/fitness/IFitness.cs
@@ -2,7 +2,7 @@ namespace GeneticProgramming
 {
     public interface IFitness
     {
-        int PassingScore { get; set;}
+        float PassingScore { get; set;}
         float CalculateScore(Genome genome);
     }
 }
